Implement filter by mime types used by activities in the objectchooser
Add a optative parameter use_mime_types to ObjectChooser constructor.
If this parameter is True, the parameter what_filter should be a
bundle_id, and the filter will be set to the collection of mime types
the activity can manage, as configured in the activity.info file.
The title in the toolbar is changed, and the what_filter combo
is not visible when this parameter is used.

A example of use is:

chooser = ObjectChooser(parent=self, what_filter=self.get_bundle_id(),
                        use_mime_types=True)

This patch need a change in sugar-toolkit-gtk3 because a paramter was
added.

Signed-off-by: Gonzalo Odiard <gonzalo@laptop.org>