Support webservice repository
This patch is a refactoring of the accounts manager for webservices to
enable services to be installed in both /usr/share/sugar/extensions
and env.get_profile_path()/extensions. There is a complexity is due to
an issue with a collision of namespace; hence a repository of loaded
modules is maintained to help disambiguate sevices installed in the
two places.

The unit test has been updated to reflect these changes.