OLPC microformat activity updater
Implement support for the widely-deployed OLPC activity microformat,
according to
http://wiki.sugarlabs.org/go/Features/Automatic_activity_updates

A new helper class was added to be able to perform partial HTTP reads.
When combined with the ZipFile class, we can download specific parts of
zip archives without having to downloda the whole archive. We do this
when we have to look up the localized activity name of a new activity.