Add webaccount/webservice extension support
This is another patch in a series of patches for interacting with
online accounts from Sugar.  The relevant feature request is [1].

This patch introduces two new sections in sugar extensions:

(1) webservice is stub for installing webservices; each web service is
installed in a subdirectory of this directory.

(2) cpsection/webaccount is a new control panel section for managing
web service accounts. Account management services are loaded into
individual subdirectories of cpsection/webaccount/services.

Note that this same directory structure can be set up in
env.get_profile_path()/extensions for web services installed by the
end user.

[1] http://wiki.sugarlabs.org/go/Features/Web_services

A unit test for the new cpsection is also included.