Add support for multiple homeviews
This patch adds support for multiple home views as per [1]. The number
of home views is determined by the list of FAVORITE_ICONS found in
jarabe/desktop/favorites.py. The icons in the list are used both for
the home view selection on the Home View toolbar and the favorites
list in the Home List View. In this version of the patch, the
school-server and go-home icons are used, resulting in two home
views. Note that the go-home icon, which has no stroke elements, is
not well suited for the favorites icon selection since it is invisible
when not selected. A new icon should be designed for this purpose.

To revert to the current single home-view configuration, one need
simply set FAVORITE_ICONS with [view-radial].

[1] http://wiki.sugarlabs.org/go/Features/Multiple_home_views