Add support for multiple homeviews
This patch adds support for multiple home views as per [1]. The number
of home views is determined by the list of VIEW_ICONS found in
/desktop/sugar/desktop/view_icons. If no gconf entry is found,
defaults from jarabe/desktop/config.py are used. The icons in the list
are used both for the home view selection on the Home View toolbar and
the favorites list in the Home List View. In this version of the
patch, view-radial is used, resulting in just one home view by
default. Similarly, in the list view, the favorities icons are found
in /desktop/sugar/desktop/favorite_icons. If not favorite icons are
found, emblem-favorites is used.

[1] http://wiki.sugarlabs.org/go/Features/Multiple_home_views