httprange: use Downloader class
Increase code re-use by using the Downloader class instead of urllib2.
The advantage here is that Soup keeps around connections for a while,
so instead of re-creating TCP streams for each partial read, we just
re-use the existing ones. This results in a noticable speedup when doing
metadata lookups.

Another advantage is that we can run the main loop while we are waiting
for a response. This means that there is no need to create a separate
thread for metadata lookups.