Add the EditToolbar with select_all and select_none buttons
Implement the changes to show the edit toolbar and the operations
'select all' and 'unselect all' only.
This version use the method find_ids in datastore, doing the operation
of select_all really fast.
Now a array contain the list of all the uids in the query, then is
not needed transverse the model to do the selection. The
uid_metadata_asoc is not needed anymore in the model.

Signed-off-by: Gonzalo Odiard <gonzalo@laptop.org>