Add the EditToolbar with select_all and select_none buttons
Implement the changes to show the edit toolbar and the operations
'select all' and 'unselect all' only.
This version use the method find_ids in datastore, doing the operation
of select_all really fast.
Use a array in the model with the list of all the uids in the query, then is
not needed transverse the model to do the selection.

v2: Now the listmodel and the listview were cleaned as suggested
    then this patch was simplified.
    Fixed TRANS comment.
    Replaced __redraw hack by the use of queue_draw.

Signed-off-by: Gonzalo Odiard <gonzalo@laptop.org>