Add get_mount_icon_name helper
Refactor code that searches the first available
icon name for given a mount point and icon size.
This search is limited to ThemedIcons.

Fallback to a default icon name if cannot be
found or not a ThemedIcon.

This patch solves the issue reported by
Suraj. Volumes with GFileIcon icons caused
Journal volumes toolbar, Journal copy-to
menu items and Frame device icon to break.

Signed-off-by: Martin Abente Lahaye <tch@sugarlabs.org>