Send pause/resume notifications from the window manager to the web activity
In the shell model for the activity: listen to changes in the state of
the Wnck window and emit custom signals "pause" or "resume"
accordingly.

In the websocket API: connect to the model signals and send
notifications to the web activity.  It follows the JSON-RPC protocol:
a notification is similar to a request except for the id, which is not
needed.