Allow the web activity to finalize sending a stop notification
Add a stop method to the shell model Activity.  It emits a stop signal
for the Apisocket to handle it.  If no callback was connected or the
callback doesn't return True, it closes the window.  So non-web
activities continue behaving as before.

Apisocket sends a stop notification to the web activity.  The web
activity must send back a close request for Apisocket to close the
window, thus completing the finalization.