ShellModel: set is_main_window when home_activity is None - #4602
In commit 58fa64 we forgot to set it in that case.  home_activity will
be None at least for the Journal.  The code in shell.py is not the
best, so we don't have another option than duplicating the code in
both parts of the conditional.

This fixes the Journal icon in the frame that was pulsing all the
time.