Load background section list store incrementally
As per SL #245, some cpsections take a long time to load, leaving the
user with the impression that something has gone wrong. This patch is
specific to the background section: it is a refactoring of the code to
load the list store, adding images one at a time, so the user sees
that the widget is working. A busy cursor is also loaded.