View Source: display toolkit source again - SL #4605
In case of a GTK activity, it will display the source of
sugar-toolkit-gtk3 .  We can enhace this later, if we could check if
the activity is GTK2 or GTK3.  There is a FIXME note in the code for
this.

In case of a web activity, it will display sugar_web inside the
activity bundle.  The initial selected file will be env.js, which is
analog to the initial file env.py in sugar-toolkit-gtk3 .