Updater: allow re-check after checking
When opening the activity update control panel, closing it once the list
of available updates has been presented, and then trying to open it again,
we would be in STATE_CHECKED and try to check for updates again
(i.e. refresh).

This was causing an UpdaterStateException even though we can and should
handle this simple refresh case.