Initialise the DBus mainloop explicitly
So that asynchronous messages and signals can be used, python-dbus must
be initialised with a main loop. This currently happens because Sugar
imports the datastore module at some point during startup.

When working on some startup-time behaviour, this caught me out.
Trying to work with dbus before that import had happened resulted in
some really strange behaviour.

Make the DBus mainloop assignment explicit to avoid this in future.