Use focus-in/out events to set the visibility in the journal - SL# 4604
The callback method of the signal 'visibility-notify-event' receive
event.get_state() == None in the case of changing from the Journal to
another activity, but not when changing to other views using F1, F2, F3.
Use focus-in/focus-out events to set the visibility of the listview,
because this works with realibility.

Signed-off-by: Gonzalo Odiard <gonzalo@laptop.org>