Avoid switch to the Journal when control panel is opened - SL #4601
If the user switch to the Journal when the Control Panel is opened
using the hotkeys or the icon in the frame, when returns to the Home,
the Contro Panel window is not show, but is focused, then the Home is useless.
This patch disable the keys F5 and F6 keys when the Control Panel
is opened.

Signed-off-by: Gonzalo Odiard <gonzalo@laptop.org>