Avoid switch to the Journal or change views when control panel is opened - SL #4601
If the user switch to the Journal when the Control Panel is opened
using the hotkeys or the icon in the frame, when returns to the Home,
the Contro Panel window is not show, but is focused, then the Home is useless.
Trying to change views while the Control Panel is opened do not
have any effect, then is disabled.
This patch disable the keys F1 to F6 keys when the Control Panel is opened.

Signed-off-by: Gonzalo Odiard <gonzalo@laptop.org>