Set limit to open activities and open instances
At the request of OLPC AU [1] (in an effort to reduce OOM freezes)
this patch uses gconf to set a maximum number of open activities
[2]. An alert is shown if the user tries to launch more activities
than the maximum asking them to close an activity before opening a new
one. If maximum_number_of_open_activites is not set or == 0, then
there is no maximum limit applied.

Further, Some activities don't behave well if more than one instance
is open (e.g., SL #4554). This patch sets a limit on the number open
instances of an activity based on a new field in activity.info:
one_open_instance.

If and only if the maximum_instances field is present in
activity.info, is it used to set the limit of open instances.

NOTE: there is a patch to activitybundle.py in the toolkit necessary
for this patch to be used.

[1] http://wiki.sugarlabs.org/go/Features/Launch_Limits
[2] /desktop/sugar/maximum_number_of_open_activities