Add gender/age attribute to profile
As per [1], this patch adds support for gender and age attributes to
the user's profile. There are modifications to both the intro screens
(gender and age selection are included after the user selects a color)
and the AboutMe section of the Control Panel, which includes mechanisms
for adjusting gender and age. The artwork used on these screens is added
by [2].

The age is stored in GConf:

'/desktop/sugar/user/age'
'/desktop/sugar/user/birth_timestamp'

The gender is stored in GConf:

'/desktop/sugar/user/gender'

[1] http://wiki.sugarlabs.org/go/Features/About_Me
[2] https://github.com/sugarlabs/sugar-artwork/pull/23