Add UI interaction to the backend
Options are displayed to the user if requested by the backend,
and a progressbar is displayed while the backup/restore is done.

The frontend and backend are separated to make easier create alternative
implementations. The remote backend is only a template, used
to test the selection when there are more than one backend available.

Signed-off-by: Gonzalo Odiard <gonzalo@laptop.org>