Implement Activity Help feature
A new shortcut (alt-shift-h) display a window with the help content
available in Help activity. The document is based in the context,
shell view or activity active. If not content is available nothing is
shown.
For activities with help content available, a menu is added in the palette
in the activity icon at the top of the frame.

Signed-off-by: Kalpa Welivitigoda <callkalpa@gmail.com>
Signed-off-by: Gonzalo Odiard <gonzalo@laptop.org>