Add age and gender options to intro sequence
As per [1], this patch adds support for gender and age attributes to
the user's profile. This patch is restricted to modifications to the
intro screens (gender and age selection are included after the user
selects a color).  The AboutMe section of the Control Panel, which
includes mechanisms for adjusting gender and age will be in a
follow-on patch. The artwork used on these screens is added by [2].

The birth time is stored in GConf:

'/desktop/sugar/user/age'
'/desktop/sugar/user/birth_timestamp'

and in gsetting:
'/desktop/sugar/user/birth-timestamp'

The gender is stored in GConf and gsettings:

'/desktop/sugar/user/gender'

[1] http://wiki.sugarlabs.org/go/Features/About_Me
[2] sugarlabs/sugar-artwork#23