Rework the wireless control panel section to respect non-wireless connections.
Fixes #4410    https://bugs.sugarlabs.org/ticket/4410

The 'Discard network history' button is peculiar to Sugar, and
is destructive to all non-wireless connections, despite being in
the 'Wireless' section of the control panel.

Relabel this button and code variables to be specific to wireless
connections.

Recode tests for non-sugar-wireless connections to exclude
non-wireless connections.

Relabel the 'Radio' checkbox and instructions to avoid confusion
over its state and action.

Test for the availability of a wireless device before building and
showing the 'Wireless' section on each invocation of the 'Network'
panel.