Add age and gender to user profile
As per [1], this patch adds support for gender and age attributes to
the user's profile. There are modifications to both the intro screens
(gender and age selection are included after the user selects a color)
and the AboutMe section of the Control Panel, which includes
mechanisms for adjusting gender and age. The artwork used on these
screens is added by [2]. Approved by the design team [3], although it seems there will be some delay in landing that after all [4]. A change to
main.py is also included since there is now a UserProfile class.

The gender and birth timestamp are stored gsettings as:

/org/sugarlabs/user/gender
/org/sugarlabs/user/birth-timestamp

A unit test is included in a separate pull request [5].
This request replaces pull requests 170, 199, 200, and 211.

[1] http://wiki.sugarlabs.org/go/Features/About_Me
[2] sugarlabs/sugar-artwork#23
[3] http://meeting.sugarlabs.org/sugar-meeting/meetings/2014-01-07T22:18:04
[4] https://github.com/sugarlabs/sugar-artwork/pull/23#issuecomment-32059303
[5] sugarlabs#201