Rework the wireless control panel section to respect non-wireless connections
Fixes #4410  https://bugs.sugarlabs.org/ticket/4410

The 'Discard network history' button is peculiar to Sugar, and
is destructive to all non-wireless connections, despite being in
the 'Wireless' section of the control panel.

Relabel the code variables and functions to be specific to
wireless connections.

Recode the functions to exclude non-wireless connections.

Relabel the 'Radio' box instructions and 'Discard ...' button
to avoid confusion over their state and actions.

If wireless hardware is not present on the system, the 'Wireless'
section of the control panel is excluded.

The availability of wireless hardware is tested with the kernel's
sys filesystem because, with the radio disabled, wireless devices
are not exposed in the NetworkManager API and it would be
inappropriate to turn the radio on--just to see if one is present.