Exclude wireless section of control panel if there's no wireless hardware.
Remove distracting interface elements when they are not needed.
Keep a clear and simple interface for Sugar.

To reveal if any wireless drivers are available in system memory,
the /sys/module/cfg80211/holders/ directory of the kernel module
database is listed. (NetworkManager on current OLPC systems do not
list wireless devices when the radio is disabled.)