Rework wireless control panel section to respect non-wireless connections
Fixes #4410  https://bugs.sugarlabs.org/ticket/4410

The 'Discard network history' button is peculiar to Sugar, and
is destructive to all non-wireless connections, despite being in
the 'Wireless' section of the control panel.

Relabel the code variables and functions to be specific to
wireless connections.

Recode the functions to exclude non-wireless connections and
include only 802-11-wireless or 802-11-olpc-mesh types.

To minimize the chances that a manually created connection (such
as 'Sugar Ad-hoc Network 2') is falsely white-listed as one of the
Sugar internal ad hoc or olpc-mesh connections, each connection
is tested against a discrete white list and specific to the
wireless mode or connection type.

Relabel the 'Radio' box instructions and 'Discard ...' button
to avoid confusion over their state and actions.

Feature description:
https://wiki.sugarlabs.org/go/Features/Network_control_panel