Microformat updater: set cache header
wiki.laptop.org, where the pages used to update the activities are
located, usually serve old cached pages. We can avoid that
setting the header Cache-Control' = 'max-age=600',
and be sure the page is not older than 10 minutes.

A reason to set this at client side, and not solve it only
in wiki.laptop.org, is the posibility of intermnediate caches,
as provider or school server proxies, or the use of other
activities servers in particular deployments.

Signed-off-by: Gonzalo Odiard <godiard@sugarlabs.org>