Add or remove system notification buttons
Check if there is no button for the received notification,
and create a new button to include it in the activities
tray.

Check if there is a button for the cleared buffer, based
on the application name, and remove it from the activities
tray.

And with this, we already have system notifications that
follows the behaviour described in the design discussions.

Signed-off-by: Martin Abente Lahaye <tch@sugarlabs.org>