Replace journal filter combo boxes
As per Feature [1], this patch replaces the filter combo boxes on the
Journal toolbar with palettes. The raionale is that combo boxes don't
work well when the option list is large. The palette is multicolumn
(when needed), making selection much easier (especially on a touch screen).

Note: There is a corresponding patch to sugar-artwork that introduces
a change to settings (in order to enable "set_important", which
positions a label next to a button icon on a toolbutton), a new icon
(view-type), and a change to the gtk css definition for scrolled
windows (to make them compatible with palette menu items).

[1] http://wiki.sugarlabs.org/go/Features/Replace_combo_box_in_journal_search