Extend notifications model
Add a buffer to the model. This buffer stores the notifications
that are received from the DBus interface. These notifications
are indexed by application name.

Add a new signal for when a buffer is emptied, for an specific
application name.

Considering that the notifications are shown in UI elements such
as Palettes, which are generated on demand, it is necessary to
store notifications so they can be added later.

Basically, this approach is follows a publisher and subscriber
model.

Signed-off-by: Martin Abente Lahaye <tch@sugarlabs.org>