Disconnect volume monitor - Fixes # 4736
We use the same object CopyMenuBuilder to populate the CopyMenu
in the ObjectPalette, in the BatchCopyMenu and in the palette
in the copy button in the detail view.
This object have callback methods for mount/unmount device events,
and need be disconnected.
This patch add a method cleanup() to disconect them, and call it
in the different cases depending on the live of the other objects.

Signed-off-by: Gonzalo Odiard <godiard@sugarlabs.org>