Add control for microphone level setting
This patch adds a microphone level and mute control to the speaker
device on the Sugar Frame. It makes use of an extension to the sound
model by Emil Dudev and a new icon, media-audio-input.svg in
icons/scaleable/devices. (The extension to the sound model also
requires a change in sugar-toolkit-gtk3 sugar3/acme-volume-alsa.c and
sugar3/acme-volume-alsa.h).

See [1] and [2] for the Design Team discussion.

[1] http://wiki.sugarlabs.org/go/File:MicLevels.png
[2] http://lists.sugarlabs.org/archive/sugar-devel/2014-March/047640.html

Try 2: Some code consolidation by Martin and Gonzalo; rename
speaker.py to audio.py

Try 3: Remove unneed notification handlers that were interferring with
GNOME/Sugar interactions. (Note that ASLAMIXER and Sugar interoperate
properly with this patch.)