add labels to webservice icons
This patch adds labels to the icons for the webservices in the control
panel. The message displayed when no webservices are installed has
also been modified to be in compliance with other such messages as per
discussion with the design team.

Try2: using Glib from gi.repository instead of glib
Try3: merge commits into one patch
Try4: fix careless typo