Require Age
OLPC AU (and other deployments) are gathering statistics based on the
user's age. But the intro pages do not require that the user specify age.

This patch added age as a required field. It also adds glue to enable
gender to be required but does not force it to be required.

There is no SL ticket associated with the patch.