More robust locale matching
The locale that is set by the manufacturing data can be of the format
es_UY.UTF-8 where as the format of the locales we get from parsing
available languages is es_UY.utf8

This patch converts UTF-8 suffixes to utf8

Note that this problem has only been observed on OLPC XO images. I was
unsuccessful in convincing OLPC to change their rules for generating
manufacturing data.

Replaces PR #338