Set proper alert window when resuming from clipboard
When using the clipboard to launch an activity, e.g., opening Write
with some text, the maximum activities alert is not always displayed
in the proper window. This patch checks to see which window is open
and displays the alert in the proper window.

Note: In order to set the alert in an activity window, a dbus method
needs to be called (See the corresponding PR in sugar-toolkit-gtk3)

Fixes #4756