Don't remove activities preset filter in the object chooser
On the old version the users are given a button `clear search`
if they don't have objects of the type the activity wanted. This
is confusing as it implies that users can insert whatever they
want as an image for example.

This fix hides the button when the user does not have any of the
requested object type. It also makes the button go back to the
activities original filter