Block KeepIcon callback in ExpandedEntry
When updating ExpandedEntry widgets to render a new journal
entry, it updates the state of the KeepIcon.

If the previous entry that was rendered in the ExpandedEntry
had a different value for keep, then it will trigger a toggle
event, since both uses the sake KeepIcon widget.

This callback is processed right away, so the other widgets still
contain previous journal entry values, and since this callback
triggers a metadata update, the old journal entry values are store
in the the new journal entry metadata, and get mixed.

Fixes #4770

Signed-off-by: Martin Abente Lahaye <tch@sugarlabs.org>