Include all UTC time zones
Fixes the issue where places like Tonga have their time zones excluded. They
need the UTC+13 and +14 instead of -11 and -10, as they are on the the other
side of the international date line to the -11 and -10 time zones.

This fix is related to bug #2195 and is based of Quozl's patch [1].
TL;DR: "increase UTC timezone coverage for Kiribati and Tonga." [1]

[1]: http://lists.sugarlabs.org/archive/sugar-devel/2011-March/030497.html