Show busy on register to school server
School server registration can take a few seconds, depending on network
activity.  Worst case is when there is no school server on an adhoc
network.

Show the shell is busy.  A busy cursor does not work for touchscreens,
so an alert is used.  This is in keeping with the alert already used
when a register to school server finishes.

Test case: connect to an adhoc network, then try to register.  A busy
message will appear, then some time later a failure message will appear.