Add a viewer for Images in the view source.
How its works:
Select an image on the treeview, and It will show
If the selected file is audio/video, an icon (audio=audio-x-generic,
video=video-x-generic) will be showed.
However if you havent selected a file, a label will be showed.
If the selected file is Unknown for the view source
(no audio, text, image, video)will be show the unknown icon

Possible fix of SL#4273