Dont save jabber without user confirmation.
Right now the jabber autosave when the user press enter in entry,
with this patch, that no happens. The users needs to click accept
button in the control panel.

Removed the timeout, because we need to set it inmediatly.
Removed the change signal in entry, because, now its set when
the user click on accept and no enter in entry..

Added "apply", so, if the user click on  accept, if apply exists,
it call apply_jabber, the function get the text of the entry,
and set it to current jabber server.

We need that, because If we dont want to show
restart alert, we couldnt get the response.

Possible Fix SL#618