Show activity name in the search entry when one activity is selected
This patch complete the activity name in the filter, even for
activities not present in the favorites view.
The activity can be started pressing Enter, by default the activity
will start the last instance, if the user want start a new instance
can press Alt+Enter. This behaviour is the same in all the views.