Fix EN and AGR po files from crashing sugar
Under some -yet to be found- circumstances, Pootle is
not replacing INTEGER and EXPRESSION place holders in
po files when creating a new language from template.

This doesn't happen in all projects, only with sugar,
sugar-toolkit-gtk3, Browse, Recall and Record.

I am fixing this manually so we can make 0.104.0 release,
and figure this out for 0.104.1.

Signed-off-by: Martin Abente Lahaye <tch@sugarlabs.org>