prevent duplicate launch timers
"failed to start" will occur if an activity is started exactly 90
seconds before any previous start.  This is easy to reproduce if an
activity is repeatedly started and stopped.

This was caused by the previous launch timers.  There was no limit to
the number of launch timers that may be pending.

Delete any existing launch timer before starting a new one.  Also
removed stale FIXME for status checking; this is done in the toolkit.
(in _child_watch_cb method of activityfactory.py)

Fixes #2974.

Tested-by: James Cameron <quozl@laptop.org>
Signed-off-by: James Cameron <quozl@laptop.org>