Migrate the view help popup to WebKit2
WebKit2 is the new version of GtkWebKit that has all the cool new stuff.  It
has a slightly different api, but it is basically the same.

Migrating to WebKit2 is needed because there can only be 1 version of WebKit
per process, meaning if the help view uses the old version everything else
is forced to as well.