Search for alternative copies after deleting a bundle, fixes #4836
If the user removes a $HOME activity, sugar should fall back to
displaying a `/usr/share` copy of the same activity.  This patch
makes the bundleregistry keep a list of alternative/outdated
activities stored on the system.  When uninstalling, this list is
used to find an alternative copy.

Steps to reproduce can be found in the ticket.

Ticket URL:  <http://bugs.sugarlabs.org/ticket/4836>