Use the TreeViewInvoker
Use the TreeViewIvoker, need move the creation fo the palettes,
from the CellRendererIcons to the TreeViews.
This simplify the code, and allow create the palettes only in the moment
that is needed. In particular, in the Journal, the palette creation is expensive,
and previously was created in the moment the mouse was over the icon,
not matter if the palette was used or not. This was the origin of bugs like [1]

[1] http://bugs.sugarlabs.org/ticket/4343