Search for alternative copies after deleting a bundle, fixes #4836
If the user removes a $HOME activity, sugar should fall back to
displaying a `/usr/share` copy of the same activity.  This patch
searches through the system paths after removing an activity in
order to add it back if needed.

Steps to reproduce can be found in the ticket.

Ticket URL:  <http://bugs.sugarlabs.org/ticket/4836>