Avoid control panel performance regression - Fixes OLPC #12856
As reported here [1], the use of NetworkManager using introspection
is much slower than call it using dbus, at least for a single
function call, due to the cost of the initial import.
This patch revert to the use of dbus, but move the logic to the sugar model,
where the interfaces list can be cached.

[1] http://dev.laptop.org/ticket/12856