brightness controls: delay updates if progress changes
Add a delay to the changes triggered by the progress bar in the display
frame palette. This is necessary for the following reasons:

First, calling a external tool is slow. Second, we do not need to update
the screen brightness for every bit of increment, if the bar is still
being modified. Finally, freezing the UI is not responsive.

Signed-off-by: Martin Abente Lahaye <tch@sugarlabs.org>