brightness controls: add sugar-backlight-helper
Add a new command line tool, called sugar-backlight-helper, that
allows sugar and users to interface with backlight devices. This
tool is inspired in gnome-settings-daemon's gsd-backlight-helper:
https://git.gnome.org/browse/gnome-settings-daemon

Add a polkit action to enable the sugar shell and unprivileged
users to execute sugar-backlight-helper to modify the backlight
device which otherwise would require root access.

Examples:

$ /usr/bin/sugar-backlight-helper --get-brightness
$ pkexec /usr/bin/sugar-backlight-helper --set-brightness INT

Signed-off-by: Martin Abente Lahaye <tch@sugarlabs.org>