brightness controls: monitor backlight device for external changes
Use a GFileMonitor to track external changes done to the backlight
device. This is necessary for two cases, first, changes made to the
device from external power managers, and second,  the fact that we
can't detect XF86MonBrightnessUp and XF86MonBrightnessDown in XOs.

Make sure we do not monitor "external changes" that we trigger.

Signed-off-by: Martin Abente Lahaye <tch@sugarlabs.org>