Correctly apply valignment in journal treeview, fixes #4852
This commit fixes the sroll state resuming in the journal code.  The
state was being restored before the tree view was heigh enough for
the Gtk.ScrolledWindow to scroll to the correct position.  This patch
changes the scroll state only after the treeview has been drawn.

Ticket Link:  http://bugs.sugarlabs.org/ticket/4852