translations: add custom entry for Guarani
A group of kids from the OLPC project in Caacupe have translated both
the sugar shell and the sugar gtk3 toolkit to Guarani. They are also in
the process of translating sugar activities [1].

Therefore, it would be a big encouragment for them if we could enable
these translations in the control panel.

This patch adds a custom entry for Guarani as is not yet included in
Fedora [2].

[1] http://translate.sugarlabs.org/gn/
[2] https://bugzilla.redhat.com/show_bug.cgi?id=1207485

Signed-off-by: Martin Abente Lahaye <tch@sugarlabs.org>