Change model query in place for Journal ListView, fixes #4852
This commit changes the way the list view model updates occur.
Instead of replacing the ListModel, we now just change the result set
query.

This means that the buggy scroll position setting code is no longer
needed.  Infact, a lot of code is removed

Steps to test:

* Go to the Journal
* Open an item in the details view
* Change title or star status
* Click back
* Notice how it has updated and the scroll position preserved

Other things tested:

* Changing mountpoints
* Selections and changing mountpoints
* Selection preservation after refresh
* Changing the filter