Help window: check if another window is opened - Fixes 4801
We need check if another window is opened before open the help window.
When the help is related to a activity, is already solved
checking activity.has_shell_window(), but if is in the Home View,
we need avoid open when the My Settings window is opened.
This patch solve the last case.