Don't check window type before change view - Fixes #4795
The code removed was introduced on commit 9406c42e to avoid change
views when the My Settings window was opened, but that didn't worked,
and we added the control of modal windows on the shell on 14956762
to solve that case.
But this control have the lateral effect of block change the views
when a popup is opened. Remove this control solves the problem.